refactor: HistoryView 관련 컴포넌트 분리
